Add ability to specify name of file for user storage in configuration and move Gson object creation to factory
The ability to specify name of file in configuration helps us use different files in development and testing, so we can not rewrite data.
The developer should not create instance of object manually in constructor, so we move creation of Gson object to factory and object of Gson will be provided by DI to all classes that need it.